Refactor TUI into package
